+++ DICOM/DicomDataset.cs
@@ -608,7 +608,7 @@ namespace Dicom
             {
                 try
                 {
-                    stringValue = element.Get<string>(-1);
+                    stringValue = element.Count == 0 ? string.Empty : element.Get<string>(-1);
                     return true;
                 }
                 catch (DicomDataException)